Move MockCallAdapterState to app-test common module
